requirements.txt
PN532_I2C
RPi.GPIO
requests
flask
busio